refactor: ♻️ improve styles and support small screen
