+++ src/components/atoms/CopyText/index.js
@@ -1,0 +1,1 @@
+export * from './CopyText';
