+++ src/components/atoms/RTTable/RTTRowBlock.module.scss
@@ -1,11 +1,11 @@
 .rttBlockRow {
   position: relative;
-  width: 32.75rem;
+  width: 100%;
   height: 3.5rem;
   background: var(--theme-page-background);
   box-shadow: var(--card-shadow);
   border-radius: 0.25rem;
-  margin-bottom: 0.125rem;
+  margin-bottom: 0.25rem;
   display: flex;
   overflow: hidden;
   animation: slideIn 1.5s ease-in;
@@ -24,8 +24,6 @@
 
   transform: rotate(180deg);
   text-align: center;
-  /* letter-spacing: 0.125rem; */
-  /* animation: scrollDown 4s linear infinite; */
 }
 
 .rttBlockRow .rttrowBlockDetail {
@@ -147,14 +145,5 @@
   }
   100% {
     color: black;
-  }
-}
-
-@keyframes scrollDown {
-  from {
-    transform: translateY(-3.25rem) rotate(180deg);
-  }
-  to {
-    transform: translateY(3.25rem) rotate(180deg);
   }
 }
