+++ src/components/atoms/RTTable/RTTable.module.scss
@@ -1,25 +1,23 @@
 .rttable {
   position: relative;
-  width: 524px;
-  box-shadow: var(--card-shadow);
-  border-radius: 4px;
-  /* overflow: scroll; */
+  border-radius: 0.125rem;
+  width: 100%;
+  height: 100%;
   min-height: 20rem;
-  /* filter: contrast(0.98); */
 }
 
 .header {
-  width: inherit;
-  height: 32px;
   display: flex;
-  padding: 0 12px;
+  height: 2rem;
+  padding: 0 0.75rem;
   align-items: center;
   justify-content: space-between;
-  background: var(--theme-page-background);
-  font-weight: normal;
   font-size: var(--font-normal);
+  border-radius: 0.5rem 0.5rem 0 0;
   color: var(--theme-page-text);
-  margin-bottom: 4px;
+  background: var(--theme-page-background);
+  border-bottom: rgba(0, 0, 0, 0.2) solid 0.125rem;
+  margin-bottom: 0.25rem;
 }
 
 .button {
@@ -40,7 +38,8 @@
   color: var(--theme-page-background);
 }
 
-.rttDataList {
-  /* max-height: 100%; */
-  /* background: red; */
+@media screen and (max-width: 768px) {
+  .rttable {
+    margin-bottom: 1rem;
+  }
 }
